Fix #197 move `GetHashMac`
